% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomsfromCUI.R
\name{atomsfromCUI}
\alias{atomsfromCUI}
\title{Retrieving UMLS Atoms}
\usage{
atomsfromCUI(CUI, vocabulary = NULL, language = NULL, pageSize = NULL)
}
\arguments{
\item{CUI}{UMLS Concept Unique Identifier}

\item{vocabulary}{The default is NULL and returns all vocabularies. It takes any root source abbreviation in the UMLS. See the “Abbreviation” column for a list of UMLS source vocabulary abbreviations.
Currently filtering by one vocabulary is supported.}

\item{language}{The default is NULL and returns all languages. It takes any 3 letter language abbreviation in the UMLS, such as “ENG”,“FRE”.
Currently filtering by one language is supported and the languages supported are french (FRE) and english (ENG).}

\item{pageSize}{The default is NULL and returns 25 pages. pageSize is a number.}
}
\value{
A data frame.
}
\description{
This function help to retrieve atoms and information about atoms for a known CUI.
}
