% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUIfromAtoms.R
\name{CUIfromAtoms}
\alias{CUIfromAtoms}
\title{Retrieving information from AUI}
\usage{
CUIfromAtoms(AUI)
}
\arguments{
\item{AUI}{Atoms Unique Identifier}
}
\value{
A data frame.
}
\description{
This function help to retrieve information about a known AUI.
}
\examples{
\dontrun{
info <- CUIfromAtoms(AUI = "A8345234")
}
}
