% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUIfromId.R
\name{CUIfromId}
\alias{CUIfromId}
\title{Retrieving UMLS Concept Unique Identifier (source-asserted Id)}
\usage{
CUIfromId(Id, vocabulary)
}
\arguments{
\item{Id}{Source-asserted identifier.}

\item{vocabulary}{Any root source abbreviation in the UMLS. See the “Abbreviation” column for a list of UMLS source vocabulary abbreviations.}
}
\value{
A UMLS Concept Unique Identifier (CUI).
#' @examples
\dontrun{cui <- CUIfromId(Id = "J45", vocabulary = "ICD10")}
}
\description{
This function help to retrieve a UMLS Concept Unique Identifier (\strong{CUI}) associate with a specific source-asserted identifier (\strong{code}).
}
