% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromCUI.R
\name{fromCUI}
\alias{fromCUI}
\title{Information about CUI}
\usage{
fromCUI(TGT, CUI)
}
\arguments{
\item{TGT}{Ticket-grant ticket, obtain with the umls_pass function.}

\item{CUI}{UMLS Concept Unique Identifier.}
}
\value{
a data frame
}
\description{
This function retrieves information about a known CUI.
}
