% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyfromSource.R
\name{hierarchyfromSource}
\alias{hierarchyfromSource}
\title{Retrieving Source-Asserted hierarchy}
\usage{
hierarchyfromSource(vocabulary, Id, type = "descendants")
}
\arguments{
\item{vocabulary}{Any root source abbreviation in the UMLS. See the “Abbreviation” column for a list of UMLS source vocabulary abbreviations.}

\item{Id}{Source-asserted identifier.}

\item{type}{The type of hierarchy e.g. "parents", "children", "ancestors" or "descendants". By default, it is "descendants".}
}
\value{
A data frame
#' @examples
\dontrun{hierarchy <- hierarchyfromSource(vocabulary = "ICD10", Id = "J45", type = "children")}
}
\description{
This function provides you a way to retrieve a hierarchical information about a known source-asserted identifier.
}
