% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdfromString.R
\name{IdfromString}
\alias{IdfromString}
\title{Retrieving source-asserted identifiers}
\usage{
IdfromString(String, vocabulary, ENG = TRUE)
}
\arguments{
\item{String}{The search term.}

\item{vocabulary}{It takes any root source abbreviation in the UMLS. See the “Abbreviation” column for a list of UMLS source vocabulary abbreviations.
Currently searching for one source vocabulary at the time.}

\item{ENG}{The default is TRUE. We recommend put it to FALSE when the search term is not in English.}
}
\value{
A source-asserted identifier (codes).
#' @examples
\dontrun{Id <- IdfromString(String = "bone fracture", vocabulary = "MSH")}
}
\description{
This function help to retrieve a source-asserted identifiers (codes) associate with a search term (string) in UMLS specific source vocabulary.
}
