% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relfromCUI.R
\name{relfromCUI}
\alias{relfromCUI}
\title{Retrieving UMLS Concept Relations}
\usage{
relfromCUI(CUI, pageSize = NULL)
}
\arguments{
\item{CUI}{UMLS Concept Unique Identifier}

\item{pageSize}{The default is NULL and returns 25 pages. pageSize is a number.}
}
\value{
A data frame
}
\description{
This function help to retrieve the NLM-asserted relationships for a known CUI.
}
