% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_pass.R
\name{umls_pass}
\alias{umls_pass}
\title{Obtain TGT}
\usage{
umls_pass(apikey)
}
\arguments{
\item{apikey}{apikey gives by UMLS}
}
\description{
Function to get UMLS Ticket-Grant Ticket.
}
\examples{
\dontrun{umls_pass(apikey = "your apikey")}
}
