% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defromCUI.R
\name{defromCUI}
\alias{defromCUI}
\title{Retrieving UMLS Concept Definitions}
\usage{
defromCUI(TGT, CUI)
}
\arguments{
\item{TGT}{UMLS ticket-grant ticket.}

\item{CUI}{UMLS Concept Unique Identifier.}
}
\value{
A data frame.
}
\description{
This function help to retrieve the source-asserted definitions for a known CUI.
}
