% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromCUI.R
\name{fromCUI}
\alias{fromCUI}
\title{Information about CUI}
\usage{
fromCUI(CUI)
}
\arguments{
\item{CUI}{UMLS Concept Unique Identifier.}
}
\value{
a data frame
#' @examples
\dontrun{x <- fromCUI(CUI = "C0018689")}
}
\description{
This function retrieves information about a known CUI.
}
